feat(pie): restore feature parity for slices
