+++ packages/arcs/src/index.ts
@@ -1,4 +1,3 @@
+export * from './motion'
 export * from './types'
-export * from './useAnimatedArc'
 export * from './useArcGenerator'
-export * from './useArcsTransition'
