+++ packages/pie/src/Slices.tsx
@@ -1,54 +1,69 @@
 import React from 'react'
-import { animated } from 'react-spring'
+import { Interpolation } from 'react-spring'
+import { useTheme } from '@nivo/core'
+import { useInheritedColor } from '@nivo/colors'
 import { ArcGenerator, useArcsTransition } from '@nivo/arcs'
-import { ComputedDatum } from './types'
-
-/*
-    <g key="slices" transform={`translate(${centerX},${centerY})`}>
-        {dataWithArc.map(datumWithArc => (
-            <PieSlice<RawDatum>
-                key={datumWithArc.id}
-                datum={datumWithArc}
-                arcGenerator={arcGenerator}
-                borderWidth={borderWidth}
-                borderColor={borderColor(datumWithArc)}
-                tooltip={tooltip}
-                isInteractive={isInteractive}
-                onClick={onClick}
-                onMouseEnter={onMouseEnter}
-                onMouseMove={onMouseMove}
-                onMouseLeave={onMouseLeave}
-                setActiveId={setActiveId}
-            />
-        ))}
-    </g>
-)
-*/
+import { ComputedDatum, CompletePieSvgProps } from './types'
+import { Slice } from './Slice'
 
 interface SlicesProps<RawDatum> {
     center: [number, number]
     data: ComputedDatum<RawDatum>[]
     arcGenerator: ArcGenerator
+    borderWidth: CompletePieSvgProps<RawDatum>['borderWidth']
+    borderColor: CompletePieSvgProps<RawDatum>['borderColor']
+    isInteractive: CompletePieSvgProps<RawDatum>['isInteractive']
+    onClick?: CompletePieSvgProps<RawDatum>['onClick']
+    onMouseEnter?: CompletePieSvgProps<RawDatum>['onMouseEnter']
+    onMouseMove?: CompletePieSvgProps<RawDatum>['onMouseMove']
+    onMouseLeave?: CompletePieSvgProps<RawDatum>['onMouseLeave']
+    setActiveId: (id: null | string | number) => void
+    tooltip: CompletePieSvgProps<RawDatum>['tooltip']
 }
 
-export const Slices = <RawDatum,>({ center, data, arcGenerator }: SlicesProps<RawDatum>) => {
+export const Slices = <RawDatum,>({
+    center,
+    data,
+    arcGenerator,
+    borderWidth,
+    borderColor,
+    isInteractive,
+    onClick,
+    onMouseEnter,
+    onMouseMove,
+    onMouseLeave,
+    setActiveId,
+    tooltip,
+}: SlicesProps<RawDatum>) => {
+    const theme = useTheme()
+    const getBorderColor = useInheritedColor<ComputedDatum<RawDatum>>(borderColor, theme)
     const { transition, interpolate } = useArcsTransition<ComputedDatum<RawDatum>>(data)
 
     return (
         <g transform={`translate(${center[0]},${center[1]})`}>
             {transition((transitionProps, datum) => {
-                console.log(transitionProps, datum)
                 return (
-                    <animated.path
+                    <Slice<RawDatum>
                         key={datum.id}
-                        fill={datum.color}
-                        d={interpolate(
-                            transitionProps.startAngle,
-                            transitionProps.endAngle,
-                            transitionProps.innerRadius,
-                            transitionProps.outerRadius,
-                            arcGenerator
-                        )}
+                        datum={datum}
+                        path={
+                            interpolate(
+                                transitionProps.startAngle,
+                                transitionProps.endAngle,
+                                transitionProps.innerRadius,
+                                transitionProps.outerRadius,
+                                arcGenerator
+                            ) as Interpolation<string>
+                        }
+                        borderWidth={borderWidth}
+                        borderColor={getBorderColor(datum)}
+                        isInteractive={isInteractive}
+                        onClick={onClick}
+                        onMouseEnter={onMouseEnter}
+                        onMouseMove={onMouseMove}
+                        onMouseLeave={onMouseLeave}
+                        setActiveId={setActiveId}
+                        tooltip={tooltip}
                     />
                 )
             })}
